backend/requirements.txt
fastapi
uvicorn[standard]
opencv-python-headless
pillow
pytesseract
mediapipe
numpy
python-multipart
